query provider
some minor changes were made to the front end. basically rendering the data tables on all the client components.

the query provider from tanstack was added. modifications were made including puting it in the provider folder, adding props that will be passed to layout and changing the export name to query provider.