Update patientForm.tsx
patient form was altered to now include different parts.

drizzle insertPatient is passed as the values that are picked for the form schema.

props are assigned be used by the form.

the submit and delete functions console log the data for now and are passed into the form that will be implemented as buttons