patient table
the patient table is now able to render actual data that is added via the form. when a user adds a new patient to the database the data is then passed to the refetched data table which instanly shows the change .

also added a delete button that will be used to delete patients.