appointment form new appointment time
the appointment form was modified to now include all of the elements needed for the user to be able to submit data attributes.

the new dialog was also modified to b e able to pass the data points from the form.

the schema was modified so that there is a start time and end time column that store the values as time types