appointment status
the status is a new attribute in the columns.tsx file that is implemented within the drop down menu in the form. this status changes as the user goes throgh the different stages of the appointment.

implemented a badge that chages color based on customized variants when the user selects a different status

The appointment type was also updated to use a dropdown selection rather than a text element.