facility form and map
the map component was updated to fix the marker and zoom level of the map. height was also fixed to allow map to correctly render on the screen.

location input now has a dropdown list after user enters an address and searches. after selecting from dropdown list the lat and lon are updated which renders a map on the screen inside the form.

all zod and api values were updated to now include the address, lat and lon