search
implemented search functionality starting with the api end points, new use-search hook. passed the data into the searchmodal component and now a list renders when the user searches for the patient. when user clicks on patient they are routed to a page that is supposed to render all of their appointments.