google map
added a new google map component that stores the long and lat values similar to the OSM package before. now this has better accuracy and better controls.

for now it is only implemented in the faiclity form but it is a reusable component that can have potential use in other places of the app but most important is the accurate coordinates which were inconsistent before.